Pull in `acts_as_commentable`
This library has been abandoned and cannot be in the gemspec if we want
to install `solidus_comments` in Rails applications using Ruby 3+.

To resolve this we're pulling in the library code to `solidus_comments`.

We pulled in a fork of the original gem that already supports Ruby 3.
That source code is available here:

  https://github.com/jarednorman/acts_as_commentable

Co-authored-by: Adam Mueller <adam@super.gd>