Reformat, refactor, and reduce unused code
After pulling in the Ruby 3-compatible `acts_as_commentable` gem into
this codebase, we saw some opportunities remove the unused functionality
of the gem and only use the parts that were being actively used by
`solidus_comments`.

To test that the commentable module is working as intended, we added
model test coverage.

Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Chris Todorov <chris@super.gd>